requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncio-dgram==2.2.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.3
mcstatus==12.0.2
multidict==6.6.3
nonebot-adapter-qq==1.6.5
nonebot2==2.4.2
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygtrie==2.5.0
python-dotenv==1.1.1
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.1
